# docs/swagger.yaml
basePath: /api
definitions:
...
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel:
    properties:
      code: {}
      message:
        example: error
        type: string
      status:
        example: error
        type: string
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel:
    properties:
      errors: {}
      message:
        example: validation error
        type: string
      status:
        example: error
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel:
    properties:
      contents: {}
      kind:
        example: collection
        type: string
      total_rows:
        example: 100
        type: integer
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.CreateCategoryRequest:
    properties:
      code:
        maxLength: 3
        minLength: 3
        type: string
      description:
        maxLength: 50
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - code
    - name
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_models.CreateEntityRequest:
    properties:
      code:
        maxLength: 3
        minLength: 3
        type: string
      description:
        maxLength: 50
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - code
    - name
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.DoGetAccountBalanceResponse:
    properties:
      accountNumber:
        example: "21100100000001"
        type: string
      actualBalance:
        example: "10000"
        type: string
      availableBalance:
        example: "10000"
        type: string
      currency:
        example: IDR
        type: string
      kind:
        example: accountBalance
        type: string
      lastUpdatedAt:
        description: ISO 8601
        example: 2024-01-22T15:51:43+0700
        type: string
      pendingBalance:
        example: "10000"
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.EntityOut:
    properties:
      code:
        type: string
      description:
        type: string
      kind:
        type: string
      name:
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.GetAccountResponse:
    properties:
      accountName:
        example: John
        type: string
      accountNumber:
        example: "21100100000001"
        type: string
      actualBalance:
        example: "10000"
        type: string
      availableBalance:
        example: "10000"
        type: string
      createdAt:
        example: "2006-01-02 15:04:05"
        type: string
      currency:
        example: IDR
        type: string
      features:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletFeature'
      kind:
        example: account
        type: string
      ownerId:
        example: "5432"
        type: string
      pendingBalance:
        example: "10000"
        type: string
      status:
        example: active
        type: string
      updatedAt:
        example: "2006-01-02 15:04:05"
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateAccountRequest:
    properties:
      features:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletFeatureReq'
      isHvt:
        example: true
        type: boolean
      status:
        example: active
        type: string
    required:
    - isHvt
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletFeature:
    properties:
      allowedNegativeBalance:
        type: boolean
      balanceRangeMax:
        type: number
      balanceRangeMin:
        type: number
      negativeBalanceLimit:
        type: number
      preset:
        type: string
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletFeatureReq:
    properties:
      allowedNegativeBalance:
        type: boolean
      balanceRangeMax:
        type: string
      balanceRangeMin:
        type: string
      negativeBalanceLimit:
        type: string
      preset:
        type: string
    type: object
...
host: localhost:9567
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a go fp transaction api docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GO FP TRANSACTION API DUCUMENTATION
  version: "1.0"
paths:
...
  /v1/account-balances/:accountNumber:
    get:
      consumes:
      - application/json
      description: Get balance by account number
      parameters:
      - description: account number
        in: path
        name: accountNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.DoGetAccountBalanceResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get balance by account number
      tags:
      - Balance
...
  /v1/accounts/{accountNumber}:
    delete:
      consumes:
      - application/json
      description: Delete account's data by account number
      parameters:
      - description: account identifier
        in: path
        name: accountNumber
        required: true
        type: string
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      - example: "21100100000001"
        in: query
        name: accountNumber
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Empty response
        "400":
          description: Bad request error. This can happen if there is an error while
            get account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "404":
          description: Data not found. This can happen if data not found while get
            account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error. This can happen if there is an error
            while get account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Delete account's data
      tags:
      - Accounts
    get:
      consumes:
      - application/json
      description: Get one account detail by account number
      parameters:
      - description: account identifier
        in: path
        name: accountNumber
        required: true
        type: string
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      - example: "21100100000001"
        in: query
        name: accountNumber
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.GetAccountResponse'
        "400":
          description: Bad request error. This can happen if there is an error while
            get account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "404":
          description: Data not found. This can happen if data not found while get
            account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error. This can happen if there is an error
            while get account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get one account by account number
      tags:
      - Accounts
    patch:
      consumes:
      - application/json
      description: Update account's data by account number
      parameters:
      - description: account identifier
        in: path
        name: accountNumber
        required: true
        type: string
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      - description: A JSON object containing payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.GetAccountResponse'
        "400":
          description: Bad request error. This can happen if there is an error while
            get account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "404":
          description: Data not found. This can happen if data not found while get
            account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error. This can happen if there is an error
            while get account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Update account's data
      tags:
      - Accounts
...
  /v1/categories:
    get:
      consumes:
      - application/json
      description: Get all data category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get all data category
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create data category
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Create data category
      tags:
      - Categories
  /v1/entities:
    get:
      consumes:
      - application/json
      description: Get all data entity
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel'
            - properties:
                contents:
                  items:
                    $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.EntityOut'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get all data entity
      tags:
      - Entities
    post:
      consumes:
      - application/json
      description: Create data entity
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.CreateEntityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestTotalRowResponseModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Create data entity
      tags:
      - Entities
...
schemes:
- http
swagger: "2.0"